move PeekableBytesIO into imageprocessing file
